extract dropdown_builder from main file
